Further conversion to Double, set up data retrieval and conversion to ChartData, added chartInterval to UserPreferences

+++ DTOs/ChartDataDTO.cs
@@ -12,6 +12,6 @@ namespace DTOs
     {
         public String Date {  get; set; }
 
-        public int Value { get; set; }
+        public Double Value { get; set; }
     }
 }
